Part 2 done

+++ 2025/day5/min_max.c
@@ -1,12 +1,14 @@
+#include "ints.c"
+
 // Returns the minimum of a and b
-long long min(long long a, long long b)
+i64 min(i64 a, i64 b)
 {
 	return a < b ? a : b;
 }
 
 
 // Returns the maximum of a and b
-long long max(long long a, long long b)
+i64 max(i64 a, i64 b)
 {
 	return a > b ? a : b;
 }
